session::error submodule
